requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
openai
twilio
stripe
gunicorn
reportlab
psycopg2-binary
SQLAlchemy